refactor ingress webook
